feat : add `App.tsx` to fetch user information when application starts
